externalize genre mapping
